backend/requirements.txt
absl-py==2.1.0
asgiref==3.9.1
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
Deprecated==1.2.15
distlib==0.3.9
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dm-tree==0.1.8
docstring_parser==0.16
docx==0.2.4
et_xmlfile==2.0.0
etils==1.10.0
executing==2.1.0
filelock==3.17.0
flatbuffers==24.3.25
fonttools==4.53.1
fpdf==1.7.2
frozendict==2.4.6
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
graphviz==0.20.3
grpcio==1.67.0
h5py==3.12.1
html5lib==1.1
idna==3.10
immutabledict==4.2.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multitasking==0.0.11
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
peewee==3.17.7
pikepdf==9.4.2
Pillow==10.0.1
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==18.0.0
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2025.2
pywin32==306
pyzmq==26.2.0
reportlab==4.2.5
requests==2.32.3
rich==13.9.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==75.2.0
simple-parsing==0.1.6
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.7
tensorflow-metadata==1.16.1
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xgboost==2.1.1
yfinance==0.2.44
zipp==3.20.2
